Refactor yaml_test: split into runner and output modules
- Created runner.rs: test execution and error comparison logic
- Created output.rs: test result formatting and summary output
- Simplified mod.rs: now just orchestrates the test flow
- Better separation of concerns for maintainability
