+++ cli/src/command/yaml_test/mod.rs
@@ -1,6 +1,8 @@
 mod error_types;
 mod markup;
+mod output;
 mod parser;
+mod runner;
 mod sentence;
 mod yaml_file;
 
@@ -12,9 +14,7 @@ pub use yaml_file::{Config, YamlTestFile};
 
 use crate::cli::YamlTestArgs;
 use crate::shell::Shell;
-use divvun_runtime::{bundle::Bundle, modules::Input};
-use futures_util::StreamExt;
-use serde_json::Value;
+use divvun_runtime::bundle::Bundle;
 
 pub async fn yaml_test(_shell: &mut Shell, args: YamlTestArgs) -> anyhow::Result<()> {
     let yaml_file = YamlTestFile::from_file(args.yaml_file.to_str().unwrap())?;
@@ -75,30 +75,21 @@ pub async fn yaml_test(_shell: &mut Shell, args: YamlTestArgs) -> anyhow::Result
                 println!("\nTest {}: {}", i + 1, sentence.text);
                 println!("  Expected errors: {}", sentence.error_count());
                 
-                // Run through grammar checker
-                let mut pipe = bundle.create(config.clone()).await?;
-                let mut stream = pipe.forward(Input::String(sentence.text.clone())).await;
-                
-                if let Some(Ok(Input::Json(output_json))) = stream.next().await {
-                    match compare_errors(sentence, &output_json) {
-                        Ok(comparison) => {
-                            if comparison.all_matched {
-                                pass_count += 1;
-                                println!("  ✓ PASS");
-                            } else {
-                                fail_count += 1;
-                                println!("  ✗ FAIL");
-                                print_comparison(&comparison);
-                            }
-                        }
-                        Err(e) => {
+                match runner::run_test(sentence, &bundle, config.clone()).await {
+                    Ok(comparison) => {
+                        if comparison.all_matched {
+                            pass_count += 1;
+                            println!("  ✓ PASS");
+                        } else {
                             fail_count += 1;
-                            println!("  ✗ FAIL: {}", e);
+                            println!("  ✗ FAIL");
+                            output::print_comparison(&comparison);
                         }
                     }
-                } else {
-                    fail_count += 1;
-                    println!("  ✗ FAIL: No JSON output from grammar checker");
+                    Err(e) => {
+                        fail_count += 1;
+                        println!("  ✗ FAIL: {}", e);
+                    }
                 }
             }
             Err(e) => {
@@ -108,119 +99,7 @@ pub async fn yaml_test(_shell: &mut Shell, args: YamlTestArgs) -> anyhow::Result
         }
     }
     
-    println!("\n{}", "=".repeat(60));
-    println!("Summary:");
-    println!("  Passed: {}", pass_count);
-    println!("  Failed: {}", fail_count);
-    println!("  Parse errors: {}", parse_error_count);
-    println!("  Total: {}", parsed_tests.len());
+    output::print_summary(pass_count, fail_count, parse_error_count, parsed_tests.len());
     
     Ok(())
-}
-
-#[derive(Debug)]
-struct ErrorComparison {
-    all_matched: bool,
-    matched: Vec<usize>,
-    unmatched_expected: Vec<usize>,
-    unmatched_actual: Vec<usize>,
-}
-
-fn print_comparison(comparison: &ErrorComparison) {
-    if !comparison.unmatched_expected.is_empty() {
-        println!("    Expected errors not found: {:?}", comparison.unmatched_expected);
-    }
-    if !comparison.unmatched_actual.is_empty() {
-        println!("    Unexpected errors found: {:?}", comparison.unmatched_actual);
-    }
-}
-
-fn compare_errors(
-    sentence: &ErrorAnnotatedSentence,
-    output_json: &Value,
-) -> anyhow::Result<ErrorComparison> {
-    let errors = output_json["errors"]
-        .as_array()
-        .ok_or_else(|| anyhow::anyhow!("No 'errors' array in output"))?;
-    
-    let mut matched = Vec::new();
-    let mut unmatched_expected: Vec<usize> = (0..sentence.errors.len()).collect();
-    let mut unmatched_actual: Vec<usize> = (0..errors.len()).collect();
-    
-    // Try to match each expected error with an actual error
-    for (exp_idx, expected_err) in sentence.errors.iter().enumerate() {
-        let mut best_match: Option<usize> = None;
-        
-        for (act_idx, actual_err) in errors.iter().enumerate() {
-            if !unmatched_actual.contains(&act_idx) {
-                continue; // Already matched
-            }
-            
-            if errors_match(expected_err, actual_err)? {
-                best_match = Some(act_idx);
-                break;
-            }
-        }
-        
-        if let Some(act_idx) = best_match {
-            matched.push(exp_idx);
-            unmatched_expected.retain(|&x| x != exp_idx);
-            unmatched_actual.retain(|&x| x != act_idx);
-        }
-    }
-    
-    Ok(ErrorComparison {
-        all_matched: unmatched_expected.is_empty() && unmatched_actual.is_empty(),
-        matched,
-        unmatched_expected,
-        unmatched_actual,
-    })
-}
-
-fn errors_match(expected: &ErrorMarkup, actual: &Value) -> anyhow::Result<bool> {
-    // Check range (start/end)
-    let actual_start = actual["start"]
-        .as_u64()
-        .ok_or_else(|| anyhow::anyhow!("Missing 'start' field"))? as usize;
-    let actual_end = actual["end"]
-        .as_u64()
-        .ok_or_else(|| anyhow::anyhow!("Missing 'end' field"))? as usize;
-    
-    if expected.start != actual_start || expected.end != actual_end {
-        return Ok(false);
-    }
-    
-    // Check form (the error text)
-    let actual_form = actual["form"]
-        .as_str()
-        .ok_or_else(|| anyhow::anyhow!("Missing 'form' field"))?;
-    
-    let expected_form = expected.form_as_string();
-    if actual_form != expected_form {
-        return Ok(false);
-    }
-    
-    // Check if at least one expected suggestion is in the actual suggestions
-    if !expected.suggestions.is_empty() {
-        let actual_suggestions = actual["suggestions"]
-            .as_array()
-            .ok_or_else(|| anyhow::anyhow!("Missing 'suggestions' field"))?;
-        
-        let actual_suggestion_strings: Vec<String> = actual_suggestions
-            .iter()
-            .filter_map(|s| s.as_str().map(|s| s.to_string()))
-            .collect();
-        
-        let has_matching_suggestion = expected.suggestions.iter().any(|exp_sug| {
-            actual_suggestion_strings.iter().any(|act_sug| {
-                act_sug == exp_sug
-            })
-        });
-        
-        if !has_matching_suggestion {
-            return Ok(false);
-        }
-    }
-    
-    Ok(true)
 }
